removed unnecessary lines from top of init.vim
many old lines were used for debugging the plugin interpreters.
Everything uses deno now so there is no need for it

removed jedi-vim as it was causing errors with ddc and ddc is more
important/newer

installed vim-lsp and got pylsp working. ddc-omni catches the omni
completions coming from this plugin, so it works like a smooth
autocomplete now.